+++ CHANGELOG.md
@@ -10,6 +10,17 @@
 
 * None.  
 
+## 1.4.0 (2019-07-25)
+
+##### Enhancements
+
+* Add `--filter-pattern` flag that allows filtering targets and their subtrees
+  from the graphviz and image output.
+  [Brian Partridge](https://github.com/brianpartridge)
+
+##### Bug Fixes
+
+* None.
 
 ## 1.3.0 (2018-08-16)